toastr updates & fix

+++ src/app/scrumsprint/auto-selected-sprint/auto-selected-sprint.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { SprintService } from '../sprint.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-auto-selected-sprint',
@@ -7,9 +8,12 @@ import { SprintService } from '../sprint.service';
   styleUrls: ['./auto-selected-sprint.component.scss']
 })
 export class AutoSelectedSprintComponent {
-  constructor(public SprintService: SprintService) {}
+  constructor(public SprintService: SprintService,private toastr: ToastrService) {}
 
   removeFromSprint(index: number): void {
-    this.SprintService.removeSelectedStory(index);
+    if (confirm('Are you sure you want to remove the story ? This action cannot be undone.')) {
+      this.SprintService.removeSelectedStory(index);
+      this.toastr.success('Story Removed successfully!', 'Success');
+    }
   }
 }
