+++ src/app/scrumsprint/story-list/story-list.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { SprintService } from '../sprint.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-story-list',
@@ -7,9 +8,12 @@ import { SprintService } from '../sprint.service';
   styleUrls: ['./story-list.component.scss']
 })
 export class StoryListComponent {
-  constructor(public sprintService: SprintService) {}
+  constructor(public sprintService: SprintService,private toastr: ToastrService) {}
 
   deleteStory(index: number): void {
-    this.sprintService.deleteStory(index);
+    if (confirm('Are you sure you want to remove the story ? This action cannot be undone.')) {
+      this.sprintService.deleteStory(index);
+      this.toastr.success('Story Removed successfully!', 'Success');
+    }
   }
 }
